Migrate stock handling from int to decimal
Updated stock-related methods and properties to use `decimal`
instead of `int` for improved precision and support for
fractional stock values. This includes:

- Changed `ReserveStockAsync` and `UpdateStockHangfireAsync`
  methods in the `Stock` class to use `decimal` for the `value`
  parameter.
- Updated `IncrementStock` and `SetStock` methods in
  `EkomBackofficeApiController` to use `decimal` for the `stock`
  parameter.
- Modified the `Stock` property in `StockData` to use
  `decimal(18,2)` in the database schema.
- Added a migration script in `OrderRepository` to convert the
  `Stock` column in the `EkomStock` table to `decimal(18,2)`.
- Updated `StockRepository` to handle `decimal` values when
  deserializing Hangfire arguments.

These changes ensure consistency across the codebase and
database, enabling fractional stock management.

+++ Ekom/Controllers/EkomBackofficeApiController.cs
@@ -231,7 +231,7 @@ public class EkomBackofficeApiController : ControllerBase
     [HttpPatch]
     [Route("stock/{id:Guid}/value/{stock}")]
     [UmbracoUserAuthorize]
-    public async Task<IActionResult> IncrementStock(Guid id, int stock)
+    public async Task<IActionResult> IncrementStock(Guid id, decimal stock)
     {
         try
         {
@@ -254,7 +254,7 @@ public class EkomBackofficeApiController : ControllerBase
     [HttpPatch]
     [Route("stock/{id:Guid}/StoreAlias/{storeAlias}/value/{stock}")]
     [UmbracoUserAuthorize]
-    public async Task<IActionResult> IncrementStock(Guid id, string storeAlias, int stock)
+    public async Task<IActionResult> IncrementStock(Guid id, string storeAlias, decimal stock)
     {
         try
         {
@@ -277,7 +277,7 @@ public class EkomBackofficeApiController : ControllerBase
     [HttpPut]
     [Route("stock/{id:Guid}/value/{stock}")]
     [UmbracoUserAuthorize]
-    public async Task<IActionResult> SetStock(Guid id, int stock)
+    public async Task<IActionResult> SetStock(Guid id, decimal stock)
     {
         try
         {
@@ -299,7 +299,7 @@ public class EkomBackofficeApiController : ControllerBase
     [HttpPut]
     [Route("stock/{id:Guid}/StoreAlias/{storeAlias}/value/{stock}")]
     [UmbracoUserAuthorize]
-    public async Task<IActionResult> SetStock(Guid id, string storeAlias, int stock)
+    public async Task<IActionResult> SetStock(Guid id, string storeAlias, decimal stock)
     {
         try
         {
